# src/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Torrust-Actix API Debugger",
    "description": "This swagger debugger page, is used for documentation and testing of the Torrust-Actix API service. Use this page to test and see how the API responses on requests, which can be used for your website or application.",
    "version": "v4.2.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Localhost /"
    }
  ],
...
  "paths": {
...
    "/stats": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get tracker statistics",
        "description": "Returns comprehensive tracker statistics including torrents, users, peers, and connection metrics",
        "operationId": "getStats",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "started": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1234567890
                    },
                    "timestamp_run_save": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1234567890
                    },
                    "timestamp_run_timeout": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1234567890
                    },
                    "timestamp_run_console": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1234567890
                    },
                    "timestamp_run_keys_timeout": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1234567890
                    },
                    "torrents": {
                      "type": "integer",
                      "format": "int64",
                      "example": 100
                    },
                    "torrents_updates": {
                      "type": "integer",
                      "format": "int64",
                      "example": 5000
                    },
                    "users": {
                      "type": "integer",
                      "format": "int64",
                      "example": 50
                    },
                    "users_updates": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1000
                    },
                    "seeds": {
                      "type": "integer",
                      "format": "int64",
                      "example": 200
                    },
                    "peers": {
                      "type": "integer",
                      "format": "int64",
                      "example": 500
                    },
                    "completed": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1000
                    },
                    "whitelist_enabled": {
                      "type": "boolean",
                      "example": true
                    },
                    "whitelist": {
                      "type": "integer",
                      "format": "int64",
                      "example": 25
                    },
                    "whitelist_updates": {
                      "type": "integer",
                      "format": "int64",
                      "example": 100
                    },
                    "blacklist_enabled": {
                      "type": "boolean",
                      "example": true
                    },
                    "blacklist": {
                      "type": "integer",
                      "format": "int64",
                      "example": 5
                    },
                    "blacklist_updates": {
                      "type": "integer",
                      "format": "int64",
                      "example": 50
                    },
                    "keys_enabled": {
                      "type": "boolean",
                      "example": true
                    },
                    "keys": {
                      "type": "integer",
                      "format": "int64",
                      "example": 10
                    },
                    "keys_updates": {
                      "type": "integer",
                      "format": "int64",
                      "example": 25
                    },
                    "tcp4_not_found": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "tcp4_failure": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "tcp4_connections_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1000
                    },
                    "tcp4_api_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 500
                    },
                    "tcp4_announces_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 10000
                    },
                    "tcp4_scrapes_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 2000
                    },
                    "tcp6_not_found": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "tcp6_failure": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "tcp6_connections_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 100
                    },
                    "tcp6_api_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 50
                    },
                    "tcp6_announces_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1000
                    },
                    "tcp6_scrapes_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 200
                    },
                    "udp4_bad_request": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "udp4_invalid_request": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "udp4_connections_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 2000
                    },
                    "udp4_announces_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 20000
                    },
                    "udp4_scrapes_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 4000
                    },
                    "udp6_bad_request": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "udp6_invalid_request": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "udp6_connections_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 200
                    },
                    "udp6_announces_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 2000
                    },
                    "udp6_scrapes_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    },
                    "udp_queue_len": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_connections_active": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_requests_sent": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_requests_received": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_responses_sent": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_responses_received": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_timeouts": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_reconnects": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_auth_success": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "ws_auth_failed": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt4_connections_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt4_announces_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt4_offers_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt4_answers_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt4_scrapes_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt4_failure": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt6_connections_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt6_announces_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt6_offers_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt6_answers_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt6_scrapes_handled": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    },
                    "wt6_failure": {
                      "type": "integer",
                      "format": "int64",
                      "example": 0
                    }
                  }
                },
                "example": {
                  "started": 1234567890,
                  "torrents": 100,
                  "torrents_updates": 5000,
                  "users": 50,
                  "users_updates": 1000,
                  "seeds": 200,
                  "peers": 500,
                  "completed": 1000,
                  "whitelist_enabled": true,
                  "whitelist": 25,
                  "whitelist_updates": 100,
                  "blacklist_enabled": true,
                  "blacklist": 5,
                  "blacklist_updates": 50,
                  "keys_enabled": true,
                  "keys": 10,
                  "keys_updates": 25
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": [
              "read:apiKey"
            ]
          }
        ]
      }
    },
...
    "/api/torrent/{info_hash}/{completed}": {
      "post": {
        "tags": [
          "Torrent"
        ],
        "summary": "Update torrent completion count",
        "description": "Increments the completion count for a specific torrent by a specified amount",
        "operationId": "updateTorrentCompletion",
        "parameters": [
          {
            "in": "path",
            "name": "info_hash",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{40}$"
            },
            "required": true,
            "description": "40-character hexadecimal info_hash of the torrent",
            "example": "1234567890123456789012345678901234567890"
          },
          {
            "in": "path",
            "name": "completed",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "required": true,
            "description": "Number of completions to add",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Torrent completion count updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          },
          "304": {
            "description": "Torrent completion count updated (conditional response)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "info_hash updated 1234567890123456789012345678901234567890"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid info_hash format",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "invalid info_hash 1234567890123456789012345678901234567890"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "bad info_hash"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Torrent not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "unknown info_hash 1234567890123456789012345678901234567890"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": [
              "write:apiKey"
            ]
          }
        ]
      }
    },
...
    "/api/key/{key_hash}/{timeout}": {
      "post": {
        "tags": [
          "Key"
        ],
        "summary": "Post key data with a hash (in HEX).",
        "parameters": [
          {
            "in": "path",
            "name": "key_hash",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{40}$"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "timeout",
            "schema": {
              "type": "integer",
              "format": "uint64"
            },
            "required": true
          }
        ],
        "operationId": "postKey",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          },
          "304": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "key_hash updated 1234567890123456789012345678901234567890"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid key_hash",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "invalid key_hash 1234567890123456789012345678901234567890"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "bad key_hash"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "token_auth": {
        "type": "apiKey",
        "name": "token",
        "in": "query"
      }
    }
  }
}